天天完善接口

+++ src/main/java/com/devil/blog/controller/CategoryController.java
@@ -8,7 +8,6 @@ import org.springframework.web.bind.annotation.DeleteMapping;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.PostMapping;
-import org.springframework.web.bind.annotation.PutMapping;
 import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestParam;
 import org.springframework.web.bind.annotation.RestController;
@@ -25,16 +24,16 @@ public class CategoryController {
 
     @GetMapping("/api/v1/categories/{id}")
     public Category getCategory(@PathVariable("id") Integer id, @RequestParam(required = false) String recursion) {
-        if(recursion == null || recursion == "true") {
+        if(recursion != null && recursion.equals("true")) {
             return categoryService.getCategoryRecurively(id);
         }
         return categoryService.getCategory(id);
     }
 
-    @PutMapping("/api/v1/categories/{id}")
-    public boolean updateCategory(@PathVariable("id") Integer id, @RequestBody Map<String, Object> map) {
-        return categoryService.updateCategoryRecursively(id, map);
-    }
+    //@PutMapping("/api/v1/categories/{id}")
+    //public boolean updateCategory(@PathVariable("id") Integer id, @RequestBody Map<String, Object> map) {
+    //    return categoryService.updateCategoryRecursively(id, map);
+    //}
 
     @PostMapping("/api/v1/categories")
     public int insertCategory(@RequestBody Map<String, Object> map) {
@@ -42,7 +41,10 @@ public class CategoryController {
     }
 
     @DeleteMapping("/api/v1/categories/{id}")
-    public boolean deleteCategory(@PathVariable("id") Integer id) {
-        return categoryService.deleteCategoryRecursively(id);
+    public boolean deleteCategory(@PathVariable("id") Integer id, @RequestParam(required = false) String recursion) {
+        if(recursion != null && recursion.equals("true")) {
+            return categoryService.deleteCategoryRecursively(id);
+        }
+        return categoryService.deleteCategory(id);
     }
 }
